Django/Day9/oz-backend-django/pyproject.toml
[tool.poetry]
name = "oz-backend-django"
version = "0.1.0"
description = ""
authors = ["jclee5419 <jclee5419@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
djangorestframework = "^3.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
